TW-2283: refactor
